Bug fixes.
git-svn-id: https://cvet.by:8443/svn/FOnlineSource@756 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: d3f8c1f791b8832e94bda09c9fb060a5358627af

+++ Source/Item.cpp
@@ -525,7 +525,7 @@ void Item::WeapLoadHolder()
 {
     if( !AmmoPid )
         SetPropertyValue< hash >( Item::PropertyAmmoPid, Proto->Weapon_DefaultAmmoPid );
-    SetPropertyValue< uint >( Item::PropertyAmmoCount, Proto->Weapon_MaxAmmoCount );
+    SetPropertyValue< ushort >( Item::PropertyAmmoCount, (ushort) Proto->Weapon_MaxAmmoCount );
 }
 
 #ifdef FONLINE_SERVER
